+++ src/rules-factories/dependency-rule.js
@@ -1,8 +1,9 @@
 const {
+  DEPENDENCY_NODES,
+  DEFAULT_DEPENDENCY_NODES,
   ADDITIONAL_DEPENDENCY_NODES,
-  PREDEFINED_DEPENDENCY_NODES,
 } = require("../constants/settings");
-const { isArray, isString } = require("../helpers/utils");
+const { getArrayOrNull } = require("../helpers/utils");
 const { fileInfo } = require("../core/elementsInfo");
 const { dependencyInfo } = require("../core/dependencyInfo");
 
@@ -24,30 +25,28 @@ module.exports = function (ruleMeta, rule, ruleOptions = {}) {
         validateRules(context.settings, options.rules, ruleOptions.validateRules);
       }
 
-      const additionalDependencyNodesSetting = context.settings[ADDITIONAL_DEPENDENCY_NODES];
-      const additionalDependencyNodes = (
-        isArray(additionalDependencyNodesSetting) ? additionalDependencyNodesSetting : []
-      ).map((dependencyNode) => {
-        if (isString(dependencyNode)) {
-          return PREDEFINED_DEPENDENCY_NODES[dependencyNode];
-        }
-
-        return dependencyNode;
-      });
-
-      const dependencyNodes = [PREDEFINED_DEPENDENCY_NODES.import, ...additionalDependencyNodes]
+      const dependencyNodesSetting = getArrayOrNull(context.settings[DEPENDENCY_NODES]);
+      const additionalDependencyNodesSetting = getArrayOrNull(
+        context.settings[ADDITIONAL_DEPENDENCY_NODES],
+      );
+      const dependencyNodes = (dependencyNodesSetting || ["import"])
+        .map((dependencyNode) => DEFAULT_DEPENDENCY_NODES[dependencyNode])
         .flat()
         .filter(Boolean);
+      const additionalDependencyNodes = additionalDependencyNodesSetting || [];
 
-      return dependencyNodes.reduce((visitors, { selector, kind }) => {
-        visitors[selector] = (node) => {
-          const dependency = dependencyInfo(node.value, kind, context);
+      return [...dependencyNodes, ...additionalDependencyNodes].reduce(
+        (visitors, { selector, kind }) => {
+          visitors[selector] = (node) => {
+            const dependency = dependencyInfo(node.value, kind, context);
 
-          rule({ file, dependency, options, node, context });
-        };
+            rule({ file, dependency, options, node, context });
+          };
 
-        return visitors;
-      }, {});
+          return visitors;
+        },
+        {},
+      );
     },
   };
 };