+++ test/rules/one-level/external-dependency-nodes.spec.js
@@ -12,17 +12,20 @@ const { absoluteFilePath } = pathResolvers("one-level");
 
 const settings = {
   ...SETTINGS.oneLevel,
+  "boundaries/dependency-nodes": ["import"],
   parserOptions: {
     // Due to dynamic import usage
     ecmaVersion: 2020,
     sourceType: "module",
   },
 };
-const typescriptSettings = TYPESCRIPT_SETTINGS.oneLevel;
-const additionalDependencyNodesSetting = {
+const typescriptSettings = {
+  ...TYPESCRIPT_SETTINGS.oneLevel,
+  "boundaries/dependency-nodes": ["import"],
+};
+const dependencyNodesSettings = {
+  "boundaries/dependency-nodes": ["import", "export", "dynamic-import"],
   "boundaries/additional-dependency-nodes": [
-    "export",
-    "dynamic-import",
     {
       // mock('source')
       selector: "CallExpression[callee.name=mock] > Literal",
@@ -56,7 +59,7 @@ const options = [
   },
 ];
 
-// Without additional dependency nodes
+// Without redefined dependency nodes
 createRuleTester(settings).run(RULE, rule, {
   valid: [
     // Modules can export Link value from react-router-dom
@@ -86,10 +89,10 @@ createRuleTester(settings).run(RULE, rule, {
   ],
 });
 
-// With additional dependency nodes
+// With redefined dependency nodes
 createRuleTester({
   ...settings,
-  ...additionalDependencyNodesSetting,
+  ...dependencyNodesSettings,
 }).run(RULE, rule, {
   valid: [
     // Components can export Link value from react-router-dom
@@ -182,7 +185,7 @@ createRuleTester({
   ],
 });
 
-// Typescript without additional dependency nodes
+// Typescript without redefined dependency nodes
 createRuleTester(typescriptSettings).run(RULE, rule, {
   valid: [
     // Components can export Link type from react-router-dom
@@ -229,10 +232,10 @@ createRuleTester(typescriptSettings).run(RULE, rule, {
   ],
 });
 
-// Typescript with additional dependency nodes
+// Typescript with redefined dependency nodes
 createRuleTester({
   ...typescriptSettings,
-  ...additionalDependencyNodesSetting,
+  ...dependencyNodesSettings,
 }).run(RULE, rule, {
   valid: [
     // Components can export Link value from react-router-dom